requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
apify-client>=1.7.0
pytrends>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0
